requirements.txt
Flask
gunicorn
groq
python-dotenv
requests
Pillow